feat: add AttendanceExportService for attendance summary and detail spreadsheets
feat: implement QuotationImportService for importing quotations from Excel files

fix: make 'plan' column nullable in daily_reports table

fix: adjust decimal precision for financial columns across multiple tables

fix: add missing indexes to daily_reports and contract-related tables

+++ app/Livewire/Concerns/CleanMoneyInput.php
@@ -12,7 +12,7 @@ trait CleanMoneyInput
     {
         if (is_string($value)) {
             $cleaned = str_replace('.', '', $value);
-            return is_numeric($cleaned) ? (int) $cleaned : $value;
+            return is_numeric($cleaned) ? (float) $cleaned : $value;
         }
         return $value;
     }
